Fixed Backend and Refactor front end

+++ backend/src/controllers/cartController.ts
@@ -2,23 +2,28 @@ import {Request, Response} from 'express';
 import ApiResponse from "../response/apiResponse";
 import ApiResponseMessages from "../response/apiResponseMessages";
 import {cartService} from "../di/di";
+import {validationResult} from "express-validator";
 
 
 export const fetchAllCarts = async (req: Request, res: Response) => {
     try {
         const userId = req.userId
         const data = await cartService.getAllProducts(userId!);
-        return ApiResponse.success(res, "Cart fetch successfully", data);
+        return ApiResponse.success(res, ApiResponseMessages.cartFetchedSuccessfully, data);
     } catch (error) {
         return ApiResponse.internalServerError(res, ApiResponseMessages.anErrorOccurred, error);
     }
 };
 export const addToCart = async (req: Request, res: Response) => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return ApiResponse.badRequest(res, ApiResponseMessages.cartNotValid, errors)
+        }
         const userId = req.userId
         const {productId} = req.body;
         const data = await cartService.addToCart(userId!, productId);
-        return ApiResponse.success(res, "Added to cart", data);
+        return ApiResponse.success(res, ApiResponseMessages.addedToCart, data);
     } catch (error) {
         return ApiResponse.internalServerError(res, ApiResponseMessages.anErrorOccurred, error);
     }
@@ -26,14 +31,18 @@ export const addToCart = async (req: Request, res: Response) => {
 
 export const updateCart = async (req: Request, res: Response) => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return ApiResponse.badRequest(res, ApiResponseMessages.cartNotValid, errors)
+        }
         const userId = req.userId
         const {productId, quantity} = req.body;
         if (quantity === 0) {
             const data = await cartService.deleteFromCart(userId!, productId);
-            return ApiResponse.success(res, "Deleted from cart", data);
+            return ApiResponse.success(res, ApiResponseMessages.deletedFromCart, data);
         } else {
             const data = await cartService.updateQuantity(userId!, productId, quantity);
-            return ApiResponse.success(res, "Cart Updated", data);
+            return ApiResponse.success(res, ApiResponseMessages.cartUpdated, data);
         }
 
     } catch (error) {
@@ -42,10 +51,14 @@ export const updateCart = async (req: Request, res: Response) => {
 };
 export const deleteFromCart = async (req: Request, res: Response) => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return ApiResponse.badRequest(res, ApiResponseMessages.cartNotValid, errors)
+        }
         const userId = req.userId
         const {productId} = req.body;
         const data = await cartService.deleteFromCart(userId!, productId);
-        return ApiResponse.success(res, "Deleted from cart", data);
+        return ApiResponse.success(res, ApiResponseMessages.deletedFromCart, data);
     } catch (error) {
         return ApiResponse.internalServerError(res, ApiResponseMessages.anErrorOccurred, error);
     }
